tu104: 7 additional reference manuals
Add 7 new Turing (tu104) reference manuals. The corresponding manuals
for Volta (gv100) have previously been published, so this is just
"catching up" for Turing.

New (for tu104) files:
    dev_bus.ref.txt
    dev_display_withoffset.ref.txt
    dev_fifo.ref.txt
    dev_master.ref.txt
    dev_pbdma.ref.txt
    dev_timer.ref.txt
    dev_top.ref.txt

Note: there is a "lead chip" in each architecture. Other chips are
generally very close to the lead chip.

For Volta, gv100 is the lead chip.
For Turing, tu104 is the lead chip.

Reviewed by: Maneet Singh

Signed-off-by: John Hubbard <jhubbard@nvidia.com>