tu104: new dev_usermode.ref.txt manual
Add one more new Turing (tu104) reference manual. The corresponding
manual for Volta (gv100) has previously been published, so this is
just "catching up" for Turing.

New (for tu104) files:
    dev_usermode.ref.txt

Note: there is a "lead chip" in each architecture. Other chips are
generally very close to the lead chip.

For Volta, gv100 is the lead chip.
For Turing, tu104 is the lead chip.

Reviewed by: Maneet Singh

Signed-off-by: John Hubbard <jhubbard@nvidia.com>