tu104: MME: Macro Method Expander documents
The MME formats changed significantly for Turing. (Note that
tu104 is the lead Turing chip, which is why this is filed
under tu104.) This commit provides some information about the
key things that changed.

This is only a small part of the MME, but it is enough to get
MME up and running on Turing, assuming that one has already
got MME running on previous architectures.

Reviewed by: Maneet Singh

Signed-off-by: John Hubbard <jhubbard@nvidia.com>