memory-to-memory-format: remove redundant PIO structs
Remove the PIO (Programmed I/O) structs from all existing
memory-to-memory-format class header files, because they are unnecessary
and add to the maintenance burden. The offsets suffice for programming
these classes.

This is not applied to every class in this directory: cl0039.h is left
alone. That's because it contains some small in-line comments in the
struct, which are helpful, and yet small. It's not a maintenance burden
because it's the first version of the class, and is small.

Reviewed-by: Maneet Singh
Signed-off-by: John Hubbard <jhubbard@nvidia.com>