Turing and Ampere interrupt maps
These are spreadsheets, in .csv format. They contain the names and
interrupt numbers for Turing and Ampere chips.

This information makes it easier to figure out interrupt routing for
drivers (such as Nouveau) that bake the number directly into the code.

However, it should be noted that it's not expected that drivers bake any
of this into the code, because the information can be queried from the
GPU itself. See the NV_PTOP_DEVICE_INFO(i) array, and specifically,
INTR_ENUM, in dev_top.ref, which is, so far, published for Turing
in this repository, at:

    manuals/turing/tu104/dev_top.ref.txt

Reviewed-by: Andy Ritger
Signed-off-by: John Hubbard <jhubbard@nvidia.com>