EVO core channel classes: _HEAD_SET_CRC_CONTROL_WIDE_PIPE_CRC bit 6
Document bit 6 in the CRC control method in the EVO core channel
classes.

This bit is set to indicate that the CRC's computed at the compositor
and the output resources employ the full width of the deepened data
pipe. When set to FALSE, CRC's are calculated in a compatibility mode,
utilizing only the number of bits used in the original CRC calculation.

Reviewed-by: Andy Ritger <aritger@nvidia.com>
Signed-off-by: John Hubbard <jhubbard@nvidia.com>